Refactor: remove circ dependency checkpoints <-> chainparams
by moving CCheckpointData into chainparams.h (as per
bitcoin@f0deec572b304663f5f8e93a573f414e563b353f)
Github-Pull: #2643
Rebased-From: 50fb7797b5a58cd927e9539f8cf4e9cf21546cfb
