+++ test/lint/lint-circular-dependencies.sh
@@ -16,7 +16,6 @@ EXPECTED_CIRCULAR_DEPENDENCIES=(
     "budget/budgetmanager -> validation -> budget/budgetmanager"
     "budget/budgetmanager -> wallet/wallet -> budget/budgetmanager"
     "chain -> legacy/stakemodifier -> chain"
-    "chainparams -> checkpoints -> chainparams"
     "chainparamsbase -> util/system -> chainparamsbase"
     "consensus/params -> consensus/upgrades -> consensus/params"
     "crypter -> wallet/wallet -> crypter"
@@ -53,7 +52,6 @@ EXPECTED_CIRCULAR_DEPENDENCIES=(
     "wallet/wallet -> wallet/walletdb -> wallet/wallet"
     "wallet/wallet -> zpivchain -> wallet/wallet"
     "zpiv/zpivmodule -> zpivchain -> zpiv/zpivmodule"
-    "chain -> chainparams -> checkpoints -> chain"
     "chain -> legacy/stakemodifier -> stakeinput -> chain"
     "chain -> legacy/stakemodifier -> validation -> chain"
     "chainparamsbase -> util/system -> logging -> chainparamsbase"
@@ -74,6 +72,7 @@ EXPECTED_CIRCULAR_DEPENDENCIES=(
     "chain -> legacy/stakemodifier -> validation -> evo/specialtx -> chain"
     "chain -> legacy/stakemodifier -> validation -> validationinterface -> chain"
     "chain -> legacy/stakemodifier -> stakeinput -> txdb -> chain"
+    "chain -> legacy/stakemodifier -> validation -> checkpoints -> chain"
     "chain -> legacy/stakemodifier -> validation -> undo -> chain"
     "chain -> legacy/stakemodifier -> validation -> pow -> chain"
     "chain -> legacy/stakemodifier -> validation -> zpivchain -> chain"
